small changes to hole sizes and positions
found I made an error with hole positions and had one drilled more than one time form two sides on body.  Also made most of the bolts 5mm in size so 4.5mm holes were we had 3.3.  also the bolt was too big at the end as my barel has an 8mm inner diameter so made the end of the bolt that size keeping 9mm for the back of the bolt.